# documentation/openapi-schema.yml
openapi: 3.0.3
info:
  title: kanban-board
  description: Board for creating tasks for projects 
  version: 1.0 T
servers:
  - url: https://localhost:8000/api/v1
...
paths:
  /user: 
    get:
      tags:
        - user
      summary: Your profile
      description: 'your profile'
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'          
        '400':
          description: Invalid user    
  
    post:
      tags:
        - user
      summary: Create user
      description: Only a registered user can do this.
      operationId: createUser
      security:
        - bearerAuth: []
      requestBody:
        description: Created user object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'          

        '400':
          description: User creation failed
        '404':
          description: User not found
...
  /user/login:
    get:
      tags:
        - user
      summary: Logs user into the system
      description: ''
      operationId: loginUser
      security:
        - bearerAuth: []      
      parameters:
        - name: login
          in: query
          description: The user name for login
          required: false
          schema:
            type: string
        - name: password_hash
          in: query
          description: The password for login
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Successful authentication
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    description: Bearer Token for further authentication
                    example: eyJ0eXAwcvvcb1QiLCJhbGciOsdfsUzI1NiJ9.eyJpZCI6IjEzMzciLCJ1c2VybmFtZSI6ImJpem9uZSIswwwfcI6MTU5NDIwOTYwMCwicm9sZSI6InVzZXIifQ.ZvkYYnyM92xvxcvNW9_hSis7_x3_9rymsDcvcbvuOcc1I
        '401':
          description: Authentication error, incorrect credentials
        '500':
          description: Internal Server Error
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 10
        login:
          type: string
          example: theUser
        password_hash:
          type: string
          example: '123456c'
        email:
          type: string
          example: john@email.com
        created_at:
          type: string
          format: date-time
          description: Creation date
          example: "2021-01-30T08:30:00Z"
        updated_at:
          type: string
          format: date-time
          description: Updation date
          example: "2021-01-30T08:30:00Z"
        is_admin:
          type: boolean
          description: User Status
...
  securitySchemes:
    bearerAuth:            
      type: http
      scheme: bearer
      bearerFormat: JWT